+++ index.html
@@ -249,7 +249,7 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script src="./js/main.js type=" type="module"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 
 </html>
