+++ js/data.js
@@ -1,0 +1,32 @@
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!'
+];
+
+const PHOTO_DESCRIPTION_LIST = [
+  'Котенок забавно спит на спине.',
+  'Пушистый кот с цветком в зубах.',
+  'Полосатый охотник выглядывает из травы.',
+  'Кот на вершине мира - на холодильнике.',
+  'Зевок во всю ширину кошачьей души.',
+  'Бесконечный день в глазах сонного кота.',
+  'Кот-йог в замысловатой позе.',
+  'Мохнатый тигренок притаился в кустах.',
+  'Кот и его собственное отражение.',
+  'Кот-тучка на окне: спит, укутавшись в занавеску.',
+  'Рыжий хитрюга с мышкой в лапах.',
+  'Кот-путешественник: на крыше машины.',
+  'Нежность и ласка в объятиях кота.',
+  'Кот-звезда: после фотосессии.',
+  'Кот, который всегда падает на лапы.'
+];
+
+const NAMES = ['София', 'Полина', 'Михаил', 'Арина', 'Марсель', 'Алиса', 'Ева', 'Лидия', 'Фёдор', 'Константин', 'Марк', 'Анна', 'Ярослав', 'Артём', 'Маргарита', 'Даниил', 'Таисия', 'Антон', 'Мадина', 'Кристина', 'Лев', 'Мирослав', 'Эмилия', 'Елизавета', 'Софья'];
+
+export { MESSAGES, PHOTO_DESCRIPTION_LIST, NAMES };
